+++ src/apps/apps.cpp
@@ -1,6 +1,7 @@
 #include "apps.h"
 
 #include "launcher/registry.h"
+#include "assistant/assistant.h"
 #include "counter/counter.h"
 #include "events/events.h"
 #include "home/home.h"
@@ -19,6 +20,7 @@ void frij_register_apps(void)
     frij_registry_add(events_app());
     frij_registry_add(stopwatch_app());
     frij_registry_add(scoreboard_app());
+    frij_registry_add(assistant_app());
     frij_registry_add(counter_app());
     frij_registry_set_settings(settings_app());
 }